+++ conftest.py
@@ -1,4 +1,5 @@
 import os
+from typing import Callable
 
 import boto3
 import pytest
@@ -8,9 +9,10 @@ from xoto3.dynamodb.types import InputItem, TableResource
 from xoto3.dynamodb.utils.table import extract_key_from_item
 
 
-def make_pytest_put_fixture_for_table(table: TableResource):
+def make_pytest_put_fixture_for_table(table_callable: Callable[[], TableResource]):
     @pytest.fixture
     def put_item_fixture():
+        table = table_callable()
         keys_put = list()
 
         def _put_item(item: InputItem):
@@ -35,9 +37,9 @@ def get_integration_test_id_table():
         table = boto3.resource("dynamodb").Table(XOTO3_INTEGRATION_TEST_ID_TABLE_NAME)
         if table.name:
             return table
-    pytest.skip(reason="No integration id table was defined")
+    pytest.skip("No integration id table was defined")
     return None
 
 
-integration_test_id_table_put = make_pytest_put_fixture_for_table(get_integration_test_id_table())
+integration_test_id_table_put = make_pytest_put_fixture_for_table(get_integration_test_id_table)
 integration_test_id_table = pytest.fixture("module")(get_integration_test_id_table)